# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IntentFence Policy API",
    "version": "0.13.0",
    "description": "Fail-closed, action-bound authorization for autonomous AI agents, MCP metadata risk scanning, and x402 payment safety with ES256-signed receipts.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "contact": { "url": "https://github.com/razel369/intentfence" },
    "x-guidance": "Immediately before a consequential MCP, HTTP, A2A, or payment action, call /api/actions/authorize and execute only if the five-minute receipt verifies and its action digest still matches. Use /api/agent-risk/scan for a free metadata-only MCP tool scan."
  },
  "servers": [
    { "url": "https://agentpass-protocol.rmalka06.chatgpt.site" }
  ],
  "security": [],
...
  "paths": {
...
    "/api/preflight/verified": {
      "post": {
        "operationId": "runX402Preflight",
        "summary": "Run an AI-agent payment safety and policy preflight",
        "description": "Paid x402 policy and spend-control preflight for AI agents before a consequential action or payment. Evaluates the caller-declared action, quoted cost against its ceiling, data-retention limit, approval requirement, and approval proof; returns safe_to_proceed, needs_review, or denied with an ES256 JWS audit receipt. The IntentFence service fee is 0.005 USDC using x402 v2 exact payment on Base mainnet (eip155:8453): call once without payment to receive PAYMENT-REQUIRED, then retry with PAYMENT-SIGNATURE. Settlement proof is returned in PAYMENT-RESPONSE.",
        "tags": ["agent-policy", "payment-safety", "spend-control", "preflight", "security", "x402", "agent-commerce"],
        "security": [],
        "x-payment-info": {
          "price": { "mode": "fixed", "currency": "USD", "amount": "0.005" },
          "protocols": [
            {
              "x402": {
                "version": 2,
                "network": "eip155:8453",
                "asset": "USDC",
                "payTo": "0x833ca7dcdb6a681ddc0c15982ef0d609bceb3a5e"
              }
            }
          ]
        },
        "x-x402-price": "$0.005",
        "x-x402": {
          "version": 2,
          "scheme": "exact",
          "price": "0.005 USDC",
          "network": "eip155:8453",
          "asset": "USDC",
          "payTo": "0x833ca7dcdb6a681ddc0c15982ef0d609bceb3a5e"
        },
        "parameters": [
          {
            "name": "PAYMENT-SIGNATURE",
            "in": "header",
            "required": false,
            "description": "Base64-encoded x402 v2 payment payload returned by an x402-compatible client.",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PreflightRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Paid preflight decision after settlement",
            "headers": {
              "PAYMENT-RESPONSE": {
                "description": "Base64-encoded x402 settlement result, including on-chain transaction proof.",
                "schema": { "type": "string" }
              },
              "X-IntentFence-Request-ID": {
                "description": "IntentFence audit correlation identifier.",
                "schema": { "type": "string", "format": "uuid" }
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaidPreflightDecision" }
              }
            }
          },
          "402": {
            "description": "Payment is required or the supplied payment could not be verified",
            "headers": {
              "PAYMENT-REQUIRED": {
                "description": "Base64-encoded x402 v2 payment requirements.",
                "required": true,
                "schema": { "type": "string" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "413": { "description": "JSON body exceeds 16 KiB; a verified payment is not settled" },
          "415": { "description": "Unsupported content type; a verified payment is not settled" }
        }
      }
    },
...
    "/api/us-cpi": {
      "get": {
        "operationId": "getOfficialUsCpi",
        "summary": "Retrieve signed official U.S. CPI and core CPI data",
        "description": "Returns official Bureau of Labor Statistics headline and core CPI index values and year-over-year rates. Results use a six-hour edge cache and include an ES256 provenance receipt. Costs 0.001 USDC through x402.",
        "tags": ["official-data", "inflation", "x402", "agent-commerce"],
        "security": [],
        "x-payment-info": {
          "price": { "mode": "fixed", "currency": "USD", "amount": "0.001" },
          "protocols": [{ "x402": { "version": 2, "network": "eip155:8453", "asset": "USDC", "payTo": "0x833ca7dcdb6a681ddc0c15982ef0d609bceb3a5e" } }]
        },
        "x-x402-price": "$0.001",
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": false,
            "description": "Optional historical period in YYYY-MM format. Omit for the latest common headline/core period with a year-over-year comparison.",
            "schema": { "type": "string", "pattern": "^[0-9]{4}-(?:0[1-9]|1[0-2])$" }
          },
          {
            "name": "PAYMENT-SIGNATURE",
            "in": "header",
            "required": false,
            "description": "Base64-encoded x402 v2 payment payload. Omit on the first call to receive PAYMENT-REQUIRED.",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Official CPI data returned and the IntentFence fee settled",
            "headers": { "PAYMENT-RESPONSE": { "description": "Base64-encoded x402 settlement result.", "schema": { "type": "string" } } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UsCpiDecision" } } }
          },
          "400": { "$ref": "#/components/responses/Problem" },
          "402": { "description": "Payment is required or could not be verified" },
          "409": { "$ref": "#/components/responses/Problem" },
          "500": { "$ref": "#/components/responses/Problem" },
          "503": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "Problem": {
        "description": "Problem Details error response",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/Problem" }
          }
        }
      },
      "InvalidRequest": {
        "description": "Invalid JSON or preflight fields",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["error", "message"],
              "properties": {
                "error": { "type": "string" },
                "message": { "type": "string" }
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "Problem": {
        "type": "object",
        "additionalProperties": false,
        "required": ["type", "title", "status", "detail", "instance"],
        "properties": {
          "type": { "type": "string", "format": "uri" },
          "title": { "type": "string" },
          "status": { "type": "integer", "minimum": 400, "maximum": 599 },
          "detail": { "type": "string" },
          "instance": { "type": "string" }
        }
      },
...
      "UsCpiDecision": {
        "type": "object",
        "required": ["intentfence", "request_id", "status", "source", "period", "cpi", "summary", "checks", "receipt"],
        "properties": {
          "intentfence": { "type": "string", "const": "0.8" },
          "request_id": { "type": "string", "format": "uuid" },
          "status": { "type": "string", "const": "verified" },
          "source": { "type": "object" },
          "period": {
            "type": "object",
            "required": ["year", "month", "name"],
            "properties": { "year": { "type": "string" }, "month": { "type": "string" }, "name": { "type": "string" } }
          },
          "cpi": {
            "type": "object",
            "required": ["headline_index", "headline_yoy_percent", "core_index", "core_yoy_percent"],
            "properties": {
              "headline_index": { "type": "number" },
              "headline_yoy_percent": { "type": "number" },
              "core_index": { "type": "number" },
              "core_yoy_percent": { "type": "number" }
            }
          },
          "summary": { "type": "string" },
          "checks": { "type": "array", "items": { "type": "object" } },
          "receipt": { "type": "object" }
        }
      },
      "PreflightRequest": {
        "type": "object",
        "required": ["subject", "action"],
        "properties": {
          "subject": { "type": "string", "minLength": 1, "maxLength": 200 },
          "action": {
            "type": "object",
            "required": ["type"],
            "properties": {
              "type": { "type": "string", "minLength": 1, "maxLength": 120 },
              "resource": { "type": "string", "maxLength": 500 }
            }
          },
          "constraints": {
            "type": "object",
            "properties": {
              "currency": { "type": "string", "maxLength": 12 },
              "cost_ceiling": { "type": "number", "minimum": 0 },
              "quoted_cost": { "type": "number", "minimum": 0 },
              "data_retention_hours": { "type": "number", "minimum": 0 },
              "human_approval": {
                "type": "string",
                "enum": ["required", "optional", "not_required"]
              }
            }
          },
          "proofs": {
            "type": "array",
            "maxItems": 20,
            "items": { "type": "string", "maxLength": 200 }
          }
        }
      },
      "PreflightDecision": {
        "type": "object",
        "required": ["intentfence", "request_id", "status", "checks", "receipt"],
        "properties": {
          "intentfence": { "type": "string", "const": "0.5" },
          "request_id": { "type": "string", "format": "uuid" },
          "status": {
            "type": "string",
            "enum": ["safe_to_proceed", "needs_review", "denied"]
          },
          "checks": { "type": "array", "items": { "type": "object" } },
          "receipt": { "type": "object" }
        }
      },
      "PaidPreflightDecision": {
        "allOf": [
          { "$ref": "#/components/schemas/PreflightDecision" },
          {
            "type": "object",
            "required": ["verification_tier"],
            "properties": {
              "verification_tier": { "type": "string", "const": "x402-settled" }
            }
          }
        ]
      }
    }
  }
}